# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export type GenerateOpenaiImageBodySize =
  (typeof GenerateOpenaiImageBodySize)[keyof typeof GenerateOpenaiImageBodySize];

export const GenerateOpenaiImageBodySize = {
  "1024x1024": "1024x1024",
  "512x512": "512x512",
  "256x256": "256x256",
} as const;

export interface GenerateOpenaiImageBody {
  prompt: string;
  size?: GenerateOpenaiImageBodySize;
}

export interface GenerateOpenaiImageResponse {
  b64_json: string;
}

export interface OpenaiError {
  error: string;
}

export interface GenerateBlogBody {
  topic: string;
  keywords?: string;
  tone?: string;
  wordCount?: number;
}

export interface GenerateAiImageBody {
  prompt: string;
  style?: string;
  aspectRatio?: string;
}

export interface GenerateAiImageResponse {
  b64_json: string;
  prompt: string;
}

export interface GenerateSocialPostBody {
  topic: string;
  brand?: string;
  platforms: string[];
  tone?: string;
}

export interface SocialPlatformPost {
  platform: string;
  content: string;
  hashtags: string[];
}

export interface GenerateSocialPostResponse {
  posts: SocialPlatformPost[];
}