# lib/api-zod/src/generated/types/generateSocialPostResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SocialPlatformPost } from "./socialPlatformPost";

export interface GenerateSocialPostResponse {
  posts: SocialPlatformPost[];
}